% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{cancel_subscription}
\alias{cancel_subscription}
\title{Cancel a subscription}
\usage{
cancel_subscription(subscription_id, end_of_period = FALSE)
}
\arguments{
\item{subscription_id}{The identifier of the subscription to be canceled.}

\item{end_of_period}{If \code{FALSE} then the subscription will be canceled
immediately, if \code{TRUE} then the subscription will be canceled at the end
of the subscription period.}
}
\description{
Cancels the subscription with the provided ID on Stripe and returns the
\link{stripe_subscription} object representing the canceled subscription.
}

