% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{create_plan}
\alias{create_plan}
\title{Create a plan}
\usage{
create_plan(
  id,
  amount,
  currency,
  interval,
  nickname,
  interval_count = NULL,
  metadata = list(),
  product = NULL,
  trial_period_days = NULL
)
}
\arguments{
\item{id}{The id of plan to create}

\item{amount}{The amount of the plan in cents. So a $4.99 plan is 499}

\item{currency}{The currency for the amount}

\item{interval}{One of \code{day, week, month or year}. The frequency with
which a subscription should be billed.}

\item{interval_count}{The number of intervals (specified in the interval
property) between each subscription billing. For example,
\code{interval = month} and \code{interval_count = 3} bills every 3 months}

\item{metadata}{A set of key/value pairs that you can attach to a plan object.
It can be useful for storing additional information about the plan in a
structured format. This will be unset if you specify \code{NULL}.}

\item{trial_period_days}{Number of trial period days granted when subscribing
a customer to this plan. \code{NULL} if the plan has no trial period}

\item{name}{Display name of the plan}

\item{statement_descriptor}{An arbitrary string to be displayed on your
customer's credit card statement. This may be up to 22 characters}
}
\description{
Creates plan at Stripe and returns the \code{stripe_plan} object representing
the plan. See the \link{stripe_plan} documentation for more details on the
plan object.
}
\examples{
# Create Test Plan 1 with a monthly charge of $9.99
new_plan <- create_plan(id = "testPlan1", name = "Test Plan 1", amount = 999,
                        currency = "usd", interval = "month")

}
