% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{delete_customer}
\alias{delete_customer}
\title{Delete a customer}
\usage{
delete_customer(customer_id)
}
\arguments{
\item{id}{ID of customer to be deleted}
}
\description{
Delete specified customer from Stripe
}

