% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{delete_plan}
\alias{delete_plan}
\title{Delete a plan}
\usage{
delete_plan(plan_id)
}
\arguments{
\item{plan_id}{The id of the plan to be removed}
}
\description{
Removes the plan with the provided \code{plan_id} from your Stripe account.
If successful a message will be displayed with the \code{plan_id} of the plan
that was removed
}
\examples{
# Remove plan with plan_id testPlan
delete_plan("testPlan")
}
