% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{get_customer}
\alias{get_customer}
\title{Get customer information}
\usage{
get_customer(id)
}
\arguments{
\item{id}{ID of customer for which to rerieve information.}
}
\value{
A \code{\link{stripe_customer}} object representing the customer
}
\description{
Retrieve information on specified customer from Stripe.
}

