% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{get_plan}
\alias{get_plan}
\title{Retrieve a plan}
\usage{
get_plan(id)
}
\arguments{
\item{id}{The id of the plan to be retrieved}
}
\description{
Fetches the plan with the provided id and returns the \code{\link{stripe_plan}}
object representing the plan.
}
\examples{
# Retrieve plan with id testPlan
plan <- get_plan("testPlan")
}

