% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{get_subscription}
\alias{get_subscription}
\title{Retrieve a subscription}
\usage{
get_subscription(id)
}
\arguments{
\item{id}{The identifier of the subscription to be retrieved.}
}
\description{
Retrieves subscription with the provided ID and returns the
\link{stripe_subcription} object representing the requested subscription
}

