% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripeApp.R
\name{setApiKey}
\alias{setApiKey}
\title{Set API key}
\usage{
setApiKey(key_value)
}
\arguments{
\item{key_value}{Character string representing the API key to be used}
}
\description{
Sets API key to be used for requests to Stripe. You can find this via your
Stripe dashboard
}
