% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_portal_session.R
\docType{class}
\name{stripe_billing_portal_session}
\alias{stripe_billing_portal_session}
\title{Billing Portal Session Object}
\format{
An R6 Class object
}
\description{
The \code{stripe_checkout_session} represents your customer's session as they pay for
one-time purchases or subscriptions through Checkout or Payment Links. We recommend
creating a new Session each time your customer attempts to pay.
}
\section{Methods}{

\itemize{
 \item \code{create}: Creates a billing_portal_session on Stripe.
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-billing_portal_session-new}{\code{stripe_billing_portal_session$new()}}
\item \href{#method-billing_portal_session-create}{\code{stripe_billing_portal_session$create()}}
\item \href{#method-billing_portal_session-clone}{\code{stripe_billing_portal_session$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-billing_portal_session-new"></a>}}
\if{latex}{\out{\hypertarget{method-billing_portal_session-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_billing_portal_session$new(..., metadata = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-billing_portal_session-create"></a>}}
\if{latex}{\out{\hypertarget{method-billing_portal_session-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_billing_portal_session$create(
  customer = NULL,
  configuration = NULL,
  flow_data = NULL,
  locale = NULL,
  on_behalf_of = NULL,
  return_url = NULL,
  metadata = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-billing_portal_session-clone"></a>}}
\if{latex}{\out{\hypertarget{method-billing_portal_session-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_billing_portal_session$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
