% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout_session.R
\docType{class}
\name{stripe_checkout_session}
\alias{stripe_checkout_session}
\title{Coupon Object}
\format{
An R6 Class object
}
\description{
The \code{stripe_checkout_session} represents your customer's session as they pay for
one-time purchases or subscriptions through Checkout or Payment Links. We recommend
creating a new Session each time your customer attempts to pay.
}
\section{Methods}{

\itemize{
 \item \code{create}: Creates a checkout_session on Stripe.
 \item \code{retrieve}: Retrieves checkout_session information from Stripe based on the
 \code{id} passed to the method.
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-checkout_session-new}{\code{stripe_checkout_session$new()}}
\item \href{#method-checkout_session-create}{\code{stripe_checkout_session$create()}}
\item \href{#method-checkout_session-retrieve}{\code{stripe_checkout_session$retrieve()}}
\item \href{#method-checkout_session-clone}{\code{stripe_checkout_session$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkout_session-new"></a>}}
\if{latex}{\out{\hypertarget{method-checkout_session-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_checkout_session$new(..., metadata = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkout_session-create"></a>}}
\if{latex}{\out{\hypertarget{method-checkout_session-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_checkout_session$create(
  client_reference_id = NULL,
  customer = NULL,
  customer_email = NULL,
  currency = NULL,
  price = NULL,
  quantity = NULL,
  metadata = list(),
  mode = NULL,
  success_url = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkout_session-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-checkout_session-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_checkout_session$retrieve(checkout_session_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkout_session-clone"></a>}}
\if{latex}{\out{\hypertarget{method-checkout_session-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_checkout_session$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
