% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoice.R
\docType{data}
\name{stripe_line_item}
\alias{stripe_line_item}
\title{Stripe Line Item}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
stripe_line_item
}
\description{
Stripe Line Item
}
\keyword{datasets}
