% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\docType{class}
\name{stripe_plan}
\alias{stripe_plan}
\title{Plan Object}
\format{
An R6 Class object
}
\description{
The \code{stripe_plan} object represents the complete set of data required to
manage subscription plans at Stripe. A subscription plan contains the pricing
and billing information to be charged for differen features or products.
With different plans, customers can be charged differently for different
services or levels.
}
\section{Methods}{

\itemize{
 \item \code{create}: Creates a plan on Stripe. Required arguments are:
  \itemize{
   \item \code{id}: The id of the plan
   \item \code{name}: Display name of the plan
   \item \code{amount}: The amount of the plan in cents. So a $4.99 plan is 499
   \item \code{currency}: The currency for the amount
   \item \code{interval}: The billing interval for the plan. Should be one of
   \code{day, week, month, year}.
 }
 \item \code{retrieve}: Retrieves plan information from Stripe based on the
 \code{id} passed to the method.
 \item \code{update}: Updates plan information at Stripe. Using this method
 the plan name, statement descriptor, trial days and metadata properties can
 be updated.
 \item \code{delete}: Deletes the plan with the provided id at Stripe.
}
}

\examples{
# Create Test Plan 1 with a monthly charge of $9.99
new_plan <- stripe_plan$new()
new_plan$create(id = "testPlan1", name = "Test Plan 1", amount = 999,
                currency = "usd", interval = "month")

# Retrieve and update a plan name
my_plan <- stripe_plan$new()
my_plan$retrieve("testPlan1")
my_plan$update(name = "My Test Plan")

# Delete plan with id "testPlan1".
delete_plan <- stripe_plan$new()
delete_plan$delete("testPlan1")

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-plan-new}{\code{stripe_plan$new()}}
\item \href{#method-plan-create}{\code{stripe_plan$create()}}
\item \href{#method-plan-retrieve}{\code{stripe_plan$retrieve()}}
\item \href{#method-plan-update}{\code{stripe_plan$update()}}
\item \href{#method-plan-delete}{\code{stripe_plan$delete()}}
\item \href{#method-plan-clone}{\code{stripe_plan$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plan-new"></a>}}
\if{latex}{\out{\hypertarget{method-plan-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_plan$new(..., metadata = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plan-create"></a>}}
\if{latex}{\out{\hypertarget{method-plan-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_plan$create(
  id,
  amount,
  currency,
  interval,
  nickname,
  interval_count = NULL,
  metadata = list(),
  product = NULL,
  trial_period_days = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plan-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-plan-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_plan$retrieve(plan_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plan-update"></a>}}
\if{latex}{\out{\hypertarget{method-plan-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_plan$update(
  metadata = NULL,
  nickname = NULL,
  product = NULL,
  trial_period_days = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plan-delete"></a>}}
\if{latex}{\out{\hypertarget{method-plan-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_plan$delete(plan_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plan-clone"></a>}}
\if{latex}{\out{\hypertarget{method-plan-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_plan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
