% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripeApp.R
\name{stripe_search}
\alias{stripe_search}
\title{Search}
\usage{
stripe_search(resource, query = NULL, limit = NULL)
}
\arguments{
\item{resource}{Character string representing object type to search for}

\item{query}{Character string representing the query in the format specified at https://docs.stripe.com/search#search-query-language}

\item{limit}{Integer representing number of records to return from 1 to 100, defaults to 10 on API backend}
}
\description{
Returns database objects using the Stripe Search
}
