% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{update_plan}
\alias{update_plan}
\title{Update a plan}
\usage{
update_plan(
  id,
  metadata = NULL,
  nickname = NULL,
  product = NULL,
  trial_period_days = NULL
)
}
\arguments{
\item{id}{The id of the plan to be updated}

\item{metadata}{A set of key/value pairs that you can attach to a plan object.
It can be useful for storing additional information about the plan in a
structured format. This will be unset if you specicy \code{NULL}.}

\item{trial_period_days}{Number of trial period days granted when subscribing
a customer to this plan. \code{NULL} if the plan has no trial period}

\item{name}{Display name of the plan}

\item{statement_descriptor}{An arbitrary string to be displayed on your
customer's credit card statement. This may be up to 22 characters}
}
\description{
Updates the plan represented by the \code{id} provided and returns the updated
\code{\link{stripe_plan}} object. THe function can be used to update a plan's
name, metadata, statement descriptor and trial period.
}
\examples{
# Update the display name of plan with id proPlan to "Pro Plan"
my_plan <- update_plan("proPlan", name = "Pro Plan")
}
