% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{update_subscription}
\alias{update_subscription}
\title{Update a subscription}
\usage{
update_subscription(
  subscription_id,
  plan = NULL,
  coupon = NULL,
  prorate = NULL,
  proration_date = NULL,
  source = NULL,
  application_fee_percent = NULL,
  metadata = list(),
  tax_percent = NULL,
  trial_end = NULL
)
}
\arguments{
\item{subscription_id}{Identifier of subscription to be updated}

\item{plan}{The identifier of the plan to update the subscription to. If
omitted, the subscription will not change plans}

\item{coupon}{The code of the coupon to apply to this subscription. A coupon
applied to a subscription will only affect invoices created for that
particular subscription}

\item{prorate}{Flag determining whether to prorate switching plans during a
billing cycle}

\item{proration_date}{If set, the proration will be calculated as though the
subscription was updated at the given time. This can be used to apply exactly
the same proration that was previewed with upcoming invoice endpoint. It can
also be used to implement custom proration logic, such as prorating by day
instead of by second, by providing the time that you wish to use for
proration calculations.}

\item{trial_end}{Unix timestamp representing the end of the trial period the
customer will get before being charged for the first time. If set, trial_end
will override the default trial period of the plan the customer is being
subscribed to}
}
\description{
Updates the subscription with the provided ID. Enables changing plans,
applying discounts, prorating, and updating trial end dates.
}
