% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupon.R
\docType{class}
\name{stripe_coupon}
\alias{stripe_coupon}
\title{Coupon Object}
\format{An R6 Class object}
\usage{
stripe_coupon
}
\description{
The \code{stripe_coupon} object represents the complete set of data required
to manage coupons at Stripe. A coupon has either a \code{percent_off} or an
\code{amount_off} and \code{currency}. If you set an \code{amount_off}, that
amount will be subtracted from any invoice’s subtotal. For example, an invoice
with a subtotal of $100 will have a final total of $0 if a coupon with an
\code{amount_off} of 20000 is applied to it and an invoice with a subtotal of
$300 will have a final total of $100 if a coupon with an \code{amount_off} of
20000 is applied to it.
}
\section{Methods}{

\itemize{
 \item \code{create}: Creates a coupon on Stripe.
 \item \code{retrieve}: Retrieves coupon information from Stripe based on the
 \code{id} passed to the method.
 \item \code{update}: Updates coupon metadata at Stripe.
 \item \code{delete}: Deletes the coupon with the provided id at Stripe.
}
}

\keyword{datasets}
