% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\docType{class}
\name{stripe_customer}
\alias{stripe_customer}
\title{Customer object}
\format{An R6 object class}
\usage{
stripe_customer
}
\description{
The \code{stripe_customer} object represents the data that is required to
manage customers at Stripe. This object allows you to perform recurring
charges and track multiple charges that are associated with the same customer.
You can use this object to create, delete, and update customers.
}
\section{Methods}{

\itemize{
 \item \code{create}: Create a new customer
 \item \code{update}: Update customer information
 \item \code{retrieve}: Retrieve information on existing customer.
 \item \code{delete}: Delete a customer
}
}
\keyword{datasets}

