% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoice.R
\name{stripe_line_item}
\alias{stripe_line_item}
\title{Stripe Line Item}
\description{
Stripe Line Item

Stripe Line Item
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{stripe_line_item$new()}}
\item \href{#method-clone}{\code{stripe_line_item$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_line_item$new(..., plan = NULL, metadata = list(), period = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stripe_line_item$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
