% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\docType{class}
\name{stripe_plan}
\alias{stripe_plan}
\title{Plan Object}
\format{An R6 Class object}
\usage{
stripe_plan
}
\description{
The \code{stripe_plan} object represents the complete set of data required to
manage subscription plans at Stripe. A subscription plan contains the pricing
and billing information to be charged for differen features or products.
With different plans, customers can be charged differently for different
services or levels.
}
\section{Methods}{

\itemize{
 \item \code{create}: Creates a plan on Stripe. Required arguments are:
  \itemize{
   \item \code{id}: The id of the plan
   \item \code{name}: Display name of the plan
   \item \code{amount}: The amount of the plan in cents. So a $4.99 plan is 499
   \item \code{currency}: The currency for the amount
   \item \code{interval}: The billing interval for the plan. Should be one of
   \code{day, week, month, year}.
 }
 \item \code{retrieve}: Retrieves plan information from Stripe based on the
 \code{id} passed to the method.
 \item \code{update}: Updates plan information at Stripe. Using this method
 the plan name, statement descriptor, trial days and metadata properties can
 be updated.
 \item \code{delete}: Deletes the plan with the provided id at Stripe.
}
}
\examples{
# Create Test Plan 1 with a monthly charge of $9.99
new_plan <- stripe_plan$new()
new_plan$create(id = "testPlan1", name = "Test Plan 1", amount = 999,
                currency = "usd", interval = "month")

# Retrieve and update a plan name
my_plan <- stripe_plan$new()
my_plan$retrieve("testPlan1")
my_plan$update(name = "My Test Plan")

# Delete plan with id "testPlan1".
delete_plan <- stripe_plan$new()
delete_plan$delete("testPlan1")

}
\keyword{datasets}

