% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripeApp.R
\name{stripe_search}
\alias{stripe_search}
\title{Search}
\usage{
stripe_search(resource, query = null)
}
\arguments{
\item{resource}{Character string representing object type to search for}

\item{query}{Character string representing the query in the format specified at https://docs.stripe.com/search#search-query-language}
}
\description{
Returns database objects using the Stripe Search
}
