% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\docType{data}
\name{stripe_subscription}
\alias{stripe_subscription}
\title{Subscription Resource}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
stripe_subscription
}
\description{
This is the Stripe subscription resource.
}
\keyword{datasets}

