% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\docType{class}
\name{stripe_subscription}
\alias{stripe_subscription}
\title{Subscription object}
\format{An R6 Class object}
\usage{
stripe_subscription
}
\description{
The \code{stripe_subscription} object represents the data and methods needed
to manage customer subscriptions. Subscriptions allow you to charge a
customer's card on a recurring basis. A subscription ties a customer to a
particular plan that was previously created.
}
\section{Methods}{

\itemize{
 \item \code{create}: Creates a subscription for a customer on Stripe.
 Required arguments are:
  \itemize{
   \item \code{plan_id}: The id of the plan that the customer is subscribing to
   \item \code{customer_id}: The id of the customer that the subscription
   should be created for.
 }
 \item \code{retrieve}: Retrieves subscription information from Stripe based
 on the \code{id} passed to the method.
 \item \code{update}: Updates subscription information at Stripe. Using this method
 the plan name, statement descriptor, trial days and metadata properties can
 be updated.
 \item \code{cancel}: Cancels the subscription with the provided id at Stripe.
}
}

\keyword{datasets}
