% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripeApp.R
\docType{class}
\name{stripeApp}
\alias{stripeApp}
\title{stripeApp object}
\format{An R6 Class object}
\usage{
stripeApp
}
\description{
The \code{stripeApp} object represents the endpoint, version and API key used
for the package. You will need to use the \code{\link{setApiKey}} function to
set the API to the one you want to use.
}
\section{Object Properties}{

\itemize{
 \item \code{endpoint}: The URL of the API endpoint
 \item \code{ver}: The version of the endpoint
 \item \code{apiKey}: The API key to be used for requests
}
}

\section{Methods}{

\itemize{
 \item \code{request_url}: Generates a URL for the specified resource
}
}
\keyword{datasets}

